+++ backend/pkg/protogen/redpanda/api/console/v1alpha1/shadowlink.pb.gw.go
@@ -63,45 +63,6 @@ func local_request_ShadowLinkService_CreateShadowLink_0(ctx context.Context, mar
 	return msg, metadata, err
 }
 
-func request_ShadowLinkService_GetShadowLink_0(ctx context.Context, marshaler runtime.Marshaler, client ShadowLinkServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var (
-		protoReq GetShadowLinkRequest
-		metadata runtime.ServerMetadata
-		err      error
-	)
-	if req.Body != nil {
-		_, _ = io.Copy(io.Discard, req.Body)
-	}
-	val, ok := pathParams["name"]
-	if !ok {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "name")
-	}
-	protoReq.Name, err = runtime.String(val)
-	if err != nil {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "name", err)
-	}
-	msg, err := client.GetShadowLink(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
-	return msg, metadata, err
-}
-
-func local_request_ShadowLinkService_GetShadowLink_0(ctx context.Context, marshaler runtime.Marshaler, server ShadowLinkServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var (
-		protoReq GetShadowLinkRequest
-		metadata runtime.ServerMetadata
-		err      error
-	)
-	val, ok := pathParams["name"]
-	if !ok {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "name")
-	}
-	protoReq.Name, err = runtime.String(val)
-	if err != nil {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "name", err)
-	}
-	msg, err := server.GetShadowLink(ctx, &protoReq)
-	return msg, metadata, err
-}
-
 var filter_ShadowLinkService_ListShadowLinks_0 = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}
 
 func request_ShadowLinkService_ListShadowLinks_0(ctx context.Context, marshaler runtime.Marshaler, client ShadowLinkServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
@@ -297,26 +258,6 @@ func RegisterShadowLinkServiceHandlerServer(ctx context.Context, mux *runtime.Se
 		}
 		forward_ShadowLinkService_CreateShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
-	mux.Handle(http.MethodGet, pattern_ShadowLinkService_GetShadowLink_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
-		ctx, cancel := context.WithCancel(req.Context())
-		defer cancel()
-		var stream runtime.ServerTransportStream
-		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
-		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/redpanda.api.console.v1alpha1.ShadowLinkService/GetShadowLink", runtime.WithHTTPPathPattern("/v1alpha1/shadowlink/{name}"))
-		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		resp, md, err := local_request_ShadowLinkService_GetShadowLink_0(annotatedContext, inboundMarshaler, server, req, pathParams)
-		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
-		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
-		if err != nil {
-			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		forward_ShadowLinkService_GetShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-	})
 	mux.Handle(http.MethodGet, pattern_ShadowLinkService_ListShadowLinks_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
@@ -434,23 +375,6 @@ func RegisterShadowLinkServiceHandlerClient(ctx context.Context, mux *runtime.Se
 		}
 		forward_ShadowLinkService_CreateShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
-	mux.Handle(http.MethodGet, pattern_ShadowLinkService_GetShadowLink_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
-		ctx, cancel := context.WithCancel(req.Context())
-		defer cancel()
-		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/redpanda.api.console.v1alpha1.ShadowLinkService/GetShadowLink", runtime.WithHTTPPathPattern("/v1alpha1/shadowlink/{name}"))
-		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		resp, md, err := request_ShadowLinkService_GetShadowLink_0(annotatedContext, inboundMarshaler, client, req, pathParams)
-		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
-		if err != nil {
-			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		forward_ShadowLinkService_GetShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-	})
 	mux.Handle(http.MethodGet, pattern_ShadowLinkService_ListShadowLinks_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
@@ -507,7 +431,6 @@ func RegisterShadowLinkServiceHandlerClient(ctx context.Context, mux *runtime.Se
 
 var (
 	pattern_ShadowLinkService_CreateShadowLink_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1alpha1", "shadowlink"}, ""))
-	pattern_ShadowLinkService_GetShadowLink_0    = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1alpha1", "shadowlink", "name"}, ""))
 	pattern_ShadowLinkService_ListShadowLinks_0  = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1alpha1", "shadowlink"}, ""))
 	pattern_ShadowLinkService_UpdateShadowLink_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1alpha1", "shadowlink", "shadow_link.name"}, ""))
 	pattern_ShadowLinkService_DeleteShadowLink_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1alpha1", "shadowlink", "name"}, ""))
@@ -515,7 +438,6 @@ var (
 
 var (
 	forward_ShadowLinkService_CreateShadowLink_0 = runtime.ForwardResponseMessage
-	forward_ShadowLinkService_GetShadowLink_0    = runtime.ForwardResponseMessage
 	forward_ShadowLinkService_ListShadowLinks_0  = runtime.ForwardResponseMessage
 	forward_ShadowLinkService_UpdateShadowLink_0 = runtime.ForwardResponseMessage
 	forward_ShadowLinkService_DeleteShadowLink_0 = runtime.ForwardResponseMessage
