+++ backend/pkg/protogen/redpanda/api/dataplane/v1alpha3/dataplanev1alpha3connect/shadowlink.connect.go
@@ -48,6 +48,9 @@ const (
 	// ShadowLinkServiceGetShadowMetricsProcedure is the fully-qualified name of the ShadowLinkService's
 	// GetShadowMetrics RPC.
 	ShadowLinkServiceGetShadowMetricsProcedure = "/redpanda.api.dataplane.v1alpha3.ShadowLinkService/GetShadowMetrics"
+	// ShadowLinkServiceGetShadowLinkProcedure is the fully-qualified name of the ShadowLinkService's
+	// GetShadowLink RPC.
+	ShadowLinkServiceGetShadowLinkProcedure = "/redpanda.api.dataplane.v1alpha3.ShadowLinkService/GetShadowLink"
 )
 
 // These variables are the protoreflect.Descriptor objects for the RPCs defined in this package.
@@ -57,6 +60,7 @@ var (
 	shadowLinkServiceListShadowLinkTopicsMethodDescriptor = shadowLinkServiceServiceDescriptor.Methods().ByName("ListShadowLinkTopics")
 	shadowLinkServiceGetShadowTopicMethodDescriptor       = shadowLinkServiceServiceDescriptor.Methods().ByName("GetShadowTopic")
 	shadowLinkServiceGetShadowMetricsMethodDescriptor     = shadowLinkServiceServiceDescriptor.Methods().ByName("GetShadowMetrics")
+	shadowLinkServiceGetShadowLinkMethodDescriptor        = shadowLinkServiceServiceDescriptor.Methods().ByName("GetShadowLink")
 )
 
 // ShadowLinkServiceClient is a client for the redpanda.api.dataplane.v1alpha3.ShadowLinkService
@@ -66,6 +70,7 @@ type ShadowLinkServiceClient interface {
 	ListShadowLinkTopics(context.Context, *connect.Request[v1alpha3.ListShadowLinkTopicsRequest]) (*connect.Response[v1alpha3.ListShadowLinkTopicsResponse], error)
 	GetShadowTopic(context.Context, *connect.Request[v1alpha3.GetShadowTopicRequest]) (*connect.Response[v1alpha3.GetShadowTopicResponse], error)
 	GetShadowMetrics(context.Context, *connect.Request[v1alpha3.GetShadowMetricsRequest]) (*connect.Response[v1alpha3.GetShadowMetricsResponse], error)
+	GetShadowLink(context.Context, *connect.Request[v1alpha3.GetShadowLinkRequest]) (*connect.Response[v1alpha3.GetShadowLinkResponse], error)
 }
 
 // NewShadowLinkServiceClient constructs a client for the
@@ -103,6 +108,12 @@ func NewShadowLinkServiceClient(httpClient connect.HTTPClient, baseURL string, o
 			connect.WithSchema(shadowLinkServiceGetShadowMetricsMethodDescriptor),
 			connect.WithClientOptions(opts...),
 		),
+		getShadowLink: connect.NewClient[v1alpha3.GetShadowLinkRequest, v1alpha3.GetShadowLinkResponse](
+			httpClient,
+			baseURL+ShadowLinkServiceGetShadowLinkProcedure,
+			connect.WithSchema(shadowLinkServiceGetShadowLinkMethodDescriptor),
+			connect.WithClientOptions(opts...),
+		),
 	}
 }
 
@@ -112,6 +123,7 @@ type shadowLinkServiceClient struct {
 	listShadowLinkTopics *connect.Client[v1alpha3.ListShadowLinkTopicsRequest, v1alpha3.ListShadowLinkTopicsResponse]
 	getShadowTopic       *connect.Client[v1alpha3.GetShadowTopicRequest, v1alpha3.GetShadowTopicResponse]
 	getShadowMetrics     *connect.Client[v1alpha3.GetShadowMetricsRequest, v1alpha3.GetShadowMetricsResponse]
+	getShadowLink        *connect.Client[v1alpha3.GetShadowLinkRequest, v1alpha3.GetShadowLinkResponse]
 }
 
 // FailOver calls redpanda.api.dataplane.v1alpha3.ShadowLinkService.FailOver.
@@ -135,6 +147,11 @@ func (c *shadowLinkServiceClient) GetShadowMetrics(ctx context.Context, req *con
 	return c.getShadowMetrics.CallUnary(ctx, req)
 }
 
+// GetShadowLink calls redpanda.api.dataplane.v1alpha3.ShadowLinkService.GetShadowLink.
+func (c *shadowLinkServiceClient) GetShadowLink(ctx context.Context, req *connect.Request[v1alpha3.GetShadowLinkRequest]) (*connect.Response[v1alpha3.GetShadowLinkResponse], error) {
+	return c.getShadowLink.CallUnary(ctx, req)
+}
+
 // ShadowLinkServiceHandler is an implementation of the
 // redpanda.api.dataplane.v1alpha3.ShadowLinkService service.
 type ShadowLinkServiceHandler interface {
@@ -142,6 +159,7 @@ type ShadowLinkServiceHandler interface {
 	ListShadowLinkTopics(context.Context, *connect.Request[v1alpha3.ListShadowLinkTopicsRequest]) (*connect.Response[v1alpha3.ListShadowLinkTopicsResponse], error)
 	GetShadowTopic(context.Context, *connect.Request[v1alpha3.GetShadowTopicRequest]) (*connect.Response[v1alpha3.GetShadowTopicResponse], error)
 	GetShadowMetrics(context.Context, *connect.Request[v1alpha3.GetShadowMetricsRequest]) (*connect.Response[v1alpha3.GetShadowMetricsResponse], error)
+	GetShadowLink(context.Context, *connect.Request[v1alpha3.GetShadowLinkRequest]) (*connect.Response[v1alpha3.GetShadowLinkResponse], error)
 }
 
 // NewShadowLinkServiceHandler builds an HTTP handler from the service implementation. It returns
@@ -174,6 +192,12 @@ func NewShadowLinkServiceHandler(svc ShadowLinkServiceHandler, opts ...connect.H
 		connect.WithSchema(shadowLinkServiceGetShadowMetricsMethodDescriptor),
 		connect.WithHandlerOptions(opts...),
 	)
+	shadowLinkServiceGetShadowLinkHandler := connect.NewUnaryHandler(
+		ShadowLinkServiceGetShadowLinkProcedure,
+		svc.GetShadowLink,
+		connect.WithSchema(shadowLinkServiceGetShadowLinkMethodDescriptor),
+		connect.WithHandlerOptions(opts...),
+	)
 	return "/redpanda.api.dataplane.v1alpha3.ShadowLinkService/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		switch r.URL.Path {
 		case ShadowLinkServiceFailOverProcedure:
@@ -184,6 +208,8 @@ func NewShadowLinkServiceHandler(svc ShadowLinkServiceHandler, opts ...connect.H
 			shadowLinkServiceGetShadowTopicHandler.ServeHTTP(w, r)
 		case ShadowLinkServiceGetShadowMetricsProcedure:
 			shadowLinkServiceGetShadowMetricsHandler.ServeHTTP(w, r)
+		case ShadowLinkServiceGetShadowLinkProcedure:
+			shadowLinkServiceGetShadowLinkHandler.ServeHTTP(w, r)
 		default:
 			http.NotFound(w, r)
 		}
@@ -207,4 +233,8 @@ func (UnimplementedShadowLinkServiceHandler) GetShadowTopic(context.Context, *co
 
 func (UnimplementedShadowLinkServiceHandler) GetShadowMetrics(context.Context, *connect.Request[v1alpha3.GetShadowMetricsRequest]) (*connect.Response[v1alpha3.GetShadowMetricsResponse], error) {
 	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("redpanda.api.dataplane.v1alpha3.ShadowLinkService.GetShadowMetrics is not implemented"))
+}
+
+func (UnimplementedShadowLinkServiceHandler) GetShadowLink(context.Context, *connect.Request[v1alpha3.GetShadowLinkRequest]) (*connect.Response[v1alpha3.GetShadowLinkResponse], error) {
+	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("redpanda.api.dataplane.v1alpha3.ShadowLinkService.GetShadowLink is not implemented"))
 }
