+++ backend/pkg/protogen/redpanda/api/dataplane/v1alpha3/shadowlink.pb.gw.go
@@ -228,6 +228,45 @@ func local_request_ShadowLinkService_GetShadowMetrics_0(ctx context.Context, mar
 	return msg, metadata, err
 }
 
+func request_ShadowLinkService_GetShadowLink_0(ctx context.Context, marshaler runtime.Marshaler, client ShadowLinkServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var (
+		protoReq GetShadowLinkRequest
+		metadata runtime.ServerMetadata
+		err      error
+	)
+	if req.Body != nil {
+		_, _ = io.Copy(io.Discard, req.Body)
+	}
+	val, ok := pathParams["name"]
+	if !ok {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "name")
+	}
+	protoReq.Name, err = runtime.String(val)
+	if err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "name", err)
+	}
+	msg, err := client.GetShadowLink(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
+	return msg, metadata, err
+}
+
+func local_request_ShadowLinkService_GetShadowLink_0(ctx context.Context, marshaler runtime.Marshaler, server ShadowLinkServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var (
+		protoReq GetShadowLinkRequest
+		metadata runtime.ServerMetadata
+		err      error
+	)
+	val, ok := pathParams["name"]
+	if !ok {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "name")
+	}
+	protoReq.Name, err = runtime.String(val)
+	if err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "name", err)
+	}
+	msg, err := server.GetShadowLink(ctx, &protoReq)
+	return msg, metadata, err
+}
+
 // RegisterShadowLinkServiceHandlerServer registers the http handlers for service ShadowLinkService to "mux".
 // UnaryRPC     :call ShadowLinkServiceServer directly.
 // StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
@@ -313,6 +352,26 @@ func RegisterShadowLinkServiceHandlerServer(ctx context.Context, mux *runtime.Se
 			return
 		}
 		forward_ShadowLinkService_GetShadowMetrics_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
+	})
+	mux.Handle(http.MethodGet, pattern_ShadowLinkService_GetShadowLink_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+		ctx, cancel := context.WithCancel(req.Context())
+		defer cancel()
+		var stream runtime.ServerTransportStream
+		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
+		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
+		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/redpanda.api.dataplane.v1alpha3.ShadowLinkService/GetShadowLink", runtime.WithHTTPPathPattern("/v1alpha3/shadowlink/{name}"))
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		resp, md, err := local_request_ShadowLinkService_GetShadowLink_0(annotatedContext, inboundMarshaler, server, req, pathParams)
+		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
+		if err != nil {
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		forward_ShadowLinkService_GetShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
 
 	return nil
@@ -422,6 +481,23 @@ func RegisterShadowLinkServiceHandlerClient(ctx context.Context, mux *runtime.Se
 		}
 		forward_ShadowLinkService_GetShadowMetrics_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
+	mux.Handle(http.MethodGet, pattern_ShadowLinkService_GetShadowLink_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+		ctx, cancel := context.WithCancel(req.Context())
+		defer cancel()
+		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
+		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/redpanda.api.dataplane.v1alpha3.ShadowLinkService/GetShadowLink", runtime.WithHTTPPathPattern("/v1alpha3/shadowlink/{name}"))
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		resp, md, err := request_ShadowLinkService_GetShadowLink_0(annotatedContext, inboundMarshaler, client, req, pathParams)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
+		if err != nil {
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		forward_ShadowLinkService_GetShadowLink_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
+	})
 	return nil
 }
 
@@ -430,6 +506,7 @@ var (
 	pattern_ShadowLinkService_ListShadowLinkTopics_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"v1alpha3", "shadowlink", "shadow_link_name", "topic"}, ""))
 	pattern_ShadowLinkService_GetShadowTopic_0       = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3, 1, 0, 4, 1, 5, 4}, []string{"v1alpha3", "shadowlink", "shadow_link_name", "topic", "topic_name"}, ""))
 	pattern_ShadowLinkService_GetShadowMetrics_0     = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"v1alpha3", "shadowlink", "name", "metrics"}, ""))
+	pattern_ShadowLinkService_GetShadowLink_0        = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1alpha3", "shadowlink", "name"}, ""))
 )
 
 var (
@@ -437,4 +514,5 @@ var (
 	forward_ShadowLinkService_ListShadowLinkTopics_0 = runtime.ForwardResponseMessage
 	forward_ShadowLinkService_GetShadowTopic_0       = runtime.ForwardResponseMessage
 	forward_ShadowLinkService_GetShadowMetrics_0     = runtime.ForwardResponseMessage
+	forward_ShadowLinkService_GetShadowLink_0        = runtime.ForwardResponseMessage
 )
