+++ frontend/src/protogen/redpanda/api/dataplane/v1alpha3/shadowlink_pb.ts
@@ -2,8 +2,8 @@
 // @generated from file redpanda/api/dataplane/v1alpha3/shadowlink.proto (package redpanda.api.dataplane.v1alpha3, syntax proto3)
 /* eslint-disable */
 
-import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
-import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
+import type { GenEnum, GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
+import { enumDesc, fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
 import { file_buf_validate_validate } from "../../../../buf/validate/validate_pb";
 import { file_google_api_annotations } from "../../../../google/api/annotations_pb";
 import { file_google_api_field_behavior } from "../../../../google/api/field_behavior_pb";
@@ -11,7 +11,7 @@ import type { Timestamp } from "@bufbuild/protobuf/wkt";
 import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
 import { file_protoc_gen_openapiv2_options_annotations } from "../../../../protoc-gen-openapiv2/options/annotations_pb";
 import { file_redpanda_api_auth_v1_authorization } from "../../auth/v1/authorization_pb";
-import type { FailOverRequestSchema, ShadowTopicState } from "../../../core/admin/v2/shadow_link_pb";
+import type { FailOverRequestSchema, ShadowLinkConfigurations, ShadowLinkState, ShadowTopicState } from "../../../core/admin/v2/shadow_link_pb";
 import { file_redpanda_core_admin_v2_shadow_link } from "../../../core/admin/v2/shadow_link_pb";
 import type { Message } from "@bufbuild/protobuf";
 
@@ -19,7 +19,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file redpanda/api/dataplane/v1alpha3/shadowlink.proto.
  */
 export const file_redpanda_api_dataplane_v1alpha3_shadowlink: GenFile = /*@__PURE__*/
-  fileDesc("CjByZWRwYW5kYS9hcGkvZGF0YXBsYW5lL3YxYWxwaGEzL3NoYWRvd2xpbmsucHJvdG8SH3JlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMiPAoQRmFpbE92ZXJSZXNwb25zZRIMCgRuYW1lGAEgASgJEhoKEmZhaWxlZF9vdmVyX3RvcGljcxgCIAMoCSKSAwobTGlzdFNoYWRvd0xpbmtUb3BpY3NSZXF1ZXN0EiMKEHNoYWRvd19saW5rX25hbWUYASABKAlCCeBBArpIA8gBARJTCgZmaWx0ZXIYAiABKAsyQy5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkxpc3RTaGFkb3dMaW5rVG9waWNzUmVxdWVzdC5GaWx0ZXISnwEKCXBhZ2Vfc2l6ZRgDIAEoBUKLAZJBdTJhTGltaXQgdGhlIHBhZ2luYXRlZCByZXNwb25zZSB0byBhIG51bWJlciBvZiBpdGVtcy4gRGVmYXVsdHMgdG8gMTAwLiBVc2UgLTEgdG8gZGlzYWJsZSBwYWdpbmF0aW9uLlkAAAAAAECPQGkAAAAAAADwv7pIEBoOGOgHKP///////////wESEgoKcGFnZV90b2tlbhgEIAEoCRpDCgZGaWx0ZXISOQoTdG9waWNfbmFtZV9jb250YWlucxgBIAEoCUIcukgZchcY+QEyEl5bYS16QS1aMC05Ll9cLV0qJCKHAQoLU2hhZG93VG9waWMSEgoKdG9waWNfbmFtZRgBIAEoCRI3CgVzdGF0ZRgCIAEoDjIoLnJlZHBhbmRhLmNvcmUuYWRtaW4udjIuU2hhZG93VG9waWNTdGF0ZRIRCgl0b3RhbF9sYWcYAyABKAMSGAoQdG90YWxfcGFydGl0aW9ucxgEIAEoBSJ8ChxMaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNlEkMKDXNoYWRvd190b3BpY3MYASADKAsyLC5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLlNoYWRvd1RvcGljEhcKD25leHRfcGFnZV90b2tlbhgCIAEoCSJbChVHZXRTaGFkb3dUb3BpY1JlcXVlc3QSIwoQc2hhZG93X2xpbmtfbmFtZRgBIAEoCUIJ4EECukgDyAEBEh0KCnRvcGljX25hbWUYAiABKAlCCeBBArpIA8gBASK/AwoWR2V0U2hhZG93VG9waWNSZXNwb25zZRISCgp0b3BpY19uYW1lGAEgASgJEjcKBXN0YXRlGAIgASgOMigucmVkcGFuZGEuY29yZS5hZG1pbi52Mi5TaGFkb3dUb3BpY1N0YXRlEhEKCXRvdGFsX2xhZxgDIAEoAxIYChB0b3RhbF9wYXJ0aXRpb25zGAQgASgFElkKCnBhcnRpdGlvbnMYBSADKAsyRS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVzcG9uc2UuUGFydGl0aW9uSW5mbxrPAQoNUGFydGl0aW9uSW5mbxIUCgxwYXJ0aXRpb25faWQYASABKAUSIQoZc291cmNlX2xhc3Rfc3RhYmxlX29mZnNldBgCIAEoAxIWCg5oaWdoX3dhdGVybWFyaxgDIAEoAxIdChVzb3VyY2VfaGlnaF93YXRlcm1hcmsYBCABKAMSCwoDbGFnGAUgASgDEkEKHXNvdXJjZV9sYXN0X3VwZGF0ZWRfdGltZXN0YW1wGAYgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcCIyChdHZXRTaGFkb3dNZXRyaWNzUmVxdWVzdBIXCgRuYW1lGAEgASgJQgngQQK6SAPIAQEibQoYR2V0U2hhZG93TWV0cmljc1Jlc3BvbnNlEh8KF3RvdGFsX3RvcGljc19yZXBsaWNhdGVkGAEgASgEEhoKEmZhaWxlZF9vdmVyX3RvcGljcxgCIAEoBBIUCgxlcnJvcl90b3BpY3MYAyABKAQyhA0KEVNoYWRvd0xpbmtTZXJ2aWNlEt8CCghGYWlsT3ZlchInLnJlZHBhbmRhLmNvcmUuYWRtaW4udjIuRmFpbE92ZXJSZXF1ZXN0GjEucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5GYWlsT3ZlclJlc3BvbnNlIvYBkkG7ARIVRmFpbCBPdmVyIFNoYWRvdyBMaW5rGkFGYWlsIG92ZXIgYSBzaGFkb3cgbGluayBvciBhIHNwZWNpZmljIHNoYWRvdyB0b3BpYyB3aXRoaW4gYSBsaW5rLkpfCgMyMDISWAofRmFpbG92ZXIgYWNjZXB0ZWQgYW5kIGluaXRpYXRlZBI1CjMaMS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkZhaWxPdmVyUmVzcG9uc2WKph0ECAMQA4LT5JMCKToBKiIkL3YxYWxwaGEzL3NoYWRvd2xpbmsve25hbWV9L2ZhaWxvdmVyEqgDChRMaXN0U2hhZG93TGlua1RvcGljcxI8LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuTGlzdFNoYWRvd0xpbmtUb3BpY3NSZXF1ZXN0Gj0ucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5MaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNlIpICkkHRARIXTGlzdCBTaGFkb3cgTGluayBUb3BpY3MaZkxpc3Qgc2hhZG93IHRvcGljIHN0YXR1cyBpbmZvcm1hdGlvbiBmb3IgYSBzcGVjaWZpYyBzaGFkb3cgbGluay4gT3B0aW9uYWw6IGZpbHRlciBiYXNlZCBvbiB0b3BpYyBuYW1lLkpOCgMyMDASRwoCT0sSQQo/Gj0ucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5MaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNliqYdBAgBEAOC0+STAi8SLS92MWFscGhhMy9zaGFkb3dsaW5rL3tzaGFkb3dfbGlua19uYW1lfS90b3BpYxLxAgoOR2V0U2hhZG93VG9waWMSNi5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVxdWVzdBo3LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93VG9waWNSZXNwb25zZSLtAZJBnwESEEdldCBTaGFkb3cgVG9waWMaQVJldHJpZXZlIGRldGFpbHMgb2YgYSBzcGVjaWZpYyBzaGFkb3cgdG9waWMgd2l0aGluIGEgc2hhZG93IGxpbmsuSkgKAzIwMBJBCgJPSxI7CjkaNy5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVzcG9uc2WKph0ECAEQA4LT5JMCPBI6L3YxYWxwaGEzL3NoYWRvd2xpbmsve3NoYWRvd19saW5rX25hbWV9L3RvcGljL3t0b3BpY19uYW1lfRKdAwoQR2V0U2hhZG93TWV0cmljcxI4LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TWV0cmljc1JlcXVlc3QaOS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd01ldHJpY3NSZXNwb25zZSKTApJB3AESF0dldCBTaGFkb3cgTGluayBNZXRyaWNzGnVSZXRyaWV2ZSBtZXRyaWNzIGZvciBhIHNwZWNpZmljIHNoYWRvdyBsaW5rLCBpbmNsdWRpbmcgdG90YWwgdG9waWNzIHJlcGxpY2F0ZWQsIGZhaWxlZCBvdmVyIHRvcGljcywgYW5kIGVycm9yIHRvcGljcy5KSgoDMjAwEkMKAk9LEj0KOxo5LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TWV0cmljc1Jlc3BvbnNliqYdBAgBEAOC0+STAiUSIy92MWFscGhhMy9zaGFkb3dsaW5rL3tuYW1lfS9tZXRyaWNzGk6SQUsKGFNoYWRvdyBMaW5rcyAoRGF0YXBsYW5lKRIvRGF0YXBsYW5lIG9wZXJhdGlvbnMgZm9yIFJlZHBhbmRhIHNoYWRvdyBsaW5rcy5CvgIKI2NvbS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzQg9TaGFkb3dsaW5rUHJvdG9QAVpnZ2l0aHViLmNvbS9yZWRwYW5kYS1kYXRhL2NvbnNvbGUvYmFja2VuZC9wa2cvcHJvdG9nZW4vcmVkcGFuZGEvYXBpL2RhdGFwbGFuZS92MWFscGhhMztkYXRhcGxhbmV2MWFscGhhM6ICA1JBRKoCH1JlZHBhbmRhLkFwaS5EYXRhcGxhbmUuVjFhbHBoYTPKAh9SZWRwYW5kYVxBcGlcRGF0YXBsYW5lXFYxYWxwaGEz4gIrUmVkcGFuZGFcQXBpXERhdGFwbGFuZVxWMWFscGhhM1xHUEJNZXRhZGF0YeoCIlJlZHBhbmRhOjpBcGk6OkRhdGFwbGFuZTo6VjFhbHBoYTNiBnByb3RvMw", [file_buf_validate_validate, file_google_api_annotations, file_google_api_field_behavior, file_google_protobuf_timestamp, file_protoc_gen_openapiv2_options_annotations, file_redpanda_api_auth_v1_authorization, file_redpanda_core_admin_v2_shadow_link]);
+  fileDesc("CjByZWRwYW5kYS9hcGkvZGF0YXBsYW5lL3YxYWxwaGEzL3NoYWRvd2xpbmsucHJvdG8SH3JlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMiPAoQRmFpbE92ZXJSZXNwb25zZRIMCgRuYW1lGAEgASgJEhoKEmZhaWxlZF9vdmVyX3RvcGljcxgCIAMoCSKSAwobTGlzdFNoYWRvd0xpbmtUb3BpY3NSZXF1ZXN0EiMKEHNoYWRvd19saW5rX25hbWUYASABKAlCCeBBArpIA8gBARJTCgZmaWx0ZXIYAiABKAsyQy5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkxpc3RTaGFkb3dMaW5rVG9waWNzUmVxdWVzdC5GaWx0ZXISnwEKCXBhZ2Vfc2l6ZRgDIAEoBUKLAZJBdTJhTGltaXQgdGhlIHBhZ2luYXRlZCByZXNwb25zZSB0byBhIG51bWJlciBvZiBpdGVtcy4gRGVmYXVsdHMgdG8gMTAwLiBVc2UgLTEgdG8gZGlzYWJsZSBwYWdpbmF0aW9uLlkAAAAAAECPQGkAAAAAAADwv7pIEBoOGOgHKP///////////wESEgoKcGFnZV90b2tlbhgEIAEoCRpDCgZGaWx0ZXISOQoTdG9waWNfbmFtZV9jb250YWlucxgBIAEoCUIcukgZchcY+QEyEl5bYS16QS1aMC05Ll9cLV0qJCKHAQoLU2hhZG93VG9waWMSEgoKdG9waWNfbmFtZRgBIAEoCRI3CgVzdGF0ZRgCIAEoDjIoLnJlZHBhbmRhLmNvcmUuYWRtaW4udjIuU2hhZG93VG9waWNTdGF0ZRIRCgl0b3RhbF9sYWcYAyABKAMSGAoQdG90YWxfcGFydGl0aW9ucxgEIAEoBSJ8ChxMaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNlEkMKDXNoYWRvd190b3BpY3MYASADKAsyLC5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLlNoYWRvd1RvcGljEhcKD25leHRfcGFnZV90b2tlbhgCIAEoCSJbChVHZXRTaGFkb3dUb3BpY1JlcXVlc3QSIwoQc2hhZG93X2xpbmtfbmFtZRgBIAEoCUIJ4EECukgDyAEBEh0KCnRvcGljX25hbWUYAiABKAlCCeBBArpIA8gBASK/AwoWR2V0U2hhZG93VG9waWNSZXNwb25zZRISCgp0b3BpY19uYW1lGAEgASgJEjcKBXN0YXRlGAIgASgOMigucmVkcGFuZGEuY29yZS5hZG1pbi52Mi5TaGFkb3dUb3BpY1N0YXRlEhEKCXRvdGFsX2xhZxgDIAEoAxIYChB0b3RhbF9wYXJ0aXRpb25zGAQgASgFElkKCnBhcnRpdGlvbnMYBSADKAsyRS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVzcG9uc2UuUGFydGl0aW9uSW5mbxrPAQoNUGFydGl0aW9uSW5mbxIUCgxwYXJ0aXRpb25faWQYASABKAUSIQoZc291cmNlX2xhc3Rfc3RhYmxlX29mZnNldBgCIAEoAxIWCg5oaWdoX3dhdGVybWFyaxgDIAEoAxIdChVzb3VyY2VfaGlnaF93YXRlcm1hcmsYBCABKAMSCwoDbGFnGAUgASgDEkEKHXNvdXJjZV9sYXN0X3VwZGF0ZWRfdGltZXN0YW1wGAYgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcCIyChdHZXRTaGFkb3dNZXRyaWNzUmVxdWVzdBIXCgRuYW1lGAEgASgJQgngQQK6SAPIAQEibQoYR2V0U2hhZG93TWV0cmljc1Jlc3BvbnNlEh8KF3RvdGFsX3RvcGljc19yZXBsaWNhdGVkGAEgASgEEhoKEmZhaWxlZF9vdmVyX3RvcGljcxgCIAEoBBIUCgxlcnJvcl90b3BpY3MYAyABKAQiLwoUR2V0U2hhZG93TGlua1JlcXVlc3QSFwoEbmFtZRgBIAEoCUIJ4EECukgDyAEBIlkKFUdldFNoYWRvd0xpbmtSZXNwb25zZRJACgtzaGFkb3dfbGluaxgBIAEoCzIrLnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuU2hhZG93TGluayKUAQoUU2hhZG93TGlua1Rhc2tTdGF0dXMSDAoEbmFtZRgBIAEoCRI5CgVzdGF0ZRgCIAEoDjIqLnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuVGFza1N0YXRlEg4KBnJlYXNvbhgDIAEoCRIRCglicm9rZXJfaWQYBCABKAUSEAoIc2hhcmRfaWQYBSABKAUingIKClNoYWRvd0xpbmsSDAoEbmFtZRgBIAEoCRILCgN1aWQYAiABKAkSSAoOY29uZmlndXJhdGlvbnMYAyABKAsyMC5yZWRwYW5kYS5jb3JlLmFkbWluLnYyLlNoYWRvd0xpbmtDb25maWd1cmF0aW9ucxI2CgVzdGF0ZRgEIAEoDjInLnJlZHBhbmRhLmNvcmUuYWRtaW4udjIuU2hhZG93TGlua1N0YXRlEksKDHRhc2tzX3N0YXR1cxgFIAMoCzI1LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuU2hhZG93TGlua1Rhc2tTdGF0dXMSJgoec3luY2VkX3NoYWRvd190b3BpY19wcm9wZXJ0aWVzGAYgAygJKqoBCglUYXNrU3RhdGUSGgoWVEFTS19TVEFURV9VTlNQRUNJRklFRBAAEhUKEVRBU0tfU1RBVEVfQUNUSVZFEAESFQoRVEFTS19TVEFURV9QQVVTRUQQAhIfChtUQVNLX1NUQVRFX0xJTktfVU5BVkFJTEFCTEUQAxIaChZUQVNLX1NUQVRFX05PVF9SVU5OSU5HEAQSFgoSVEFTS19TVEFURV9GQVVMVEVEEAUyxg8KEVNoYWRvd0xpbmtTZXJ2aWNlEt8CCghGYWlsT3ZlchInLnJlZHBhbmRhLmNvcmUuYWRtaW4udjIuRmFpbE92ZXJSZXF1ZXN0GjEucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5GYWlsT3ZlclJlc3BvbnNlIvYBkkG7ARIVRmFpbCBPdmVyIFNoYWRvdyBMaW5rGkFGYWlsIG92ZXIgYSBzaGFkb3cgbGluayBvciBhIHNwZWNpZmljIHNoYWRvdyB0b3BpYyB3aXRoaW4gYSBsaW5rLkpfCgMyMDISWAofRmFpbG92ZXIgYWNjZXB0ZWQgYW5kIGluaXRpYXRlZBI1CjMaMS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkZhaWxPdmVyUmVzcG9uc2WKph0ECAMQA4LT5JMCKToBKiIkL3YxYWxwaGEzL3NoYWRvd2xpbmsve25hbWV9L2ZhaWxvdmVyEqgDChRMaXN0U2hhZG93TGlua1RvcGljcxI8LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuTGlzdFNoYWRvd0xpbmtUb3BpY3NSZXF1ZXN0Gj0ucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5MaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNlIpICkkHRARIXTGlzdCBTaGFkb3cgTGluayBUb3BpY3MaZkxpc3Qgc2hhZG93IHRvcGljIHN0YXR1cyBpbmZvcm1hdGlvbiBmb3IgYSBzcGVjaWZpYyBzaGFkb3cgbGluay4gT3B0aW9uYWw6IGZpbHRlciBiYXNlZCBvbiB0b3BpYyBuYW1lLkpOCgMyMDASRwoCT0sSQQo/Gj0ucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5MaXN0U2hhZG93TGlua1RvcGljc1Jlc3BvbnNliqYdBAgBEAOC0+STAi8SLS92MWFscGhhMy9zaGFkb3dsaW5rL3tzaGFkb3dfbGlua19uYW1lfS90b3BpYxLxAgoOR2V0U2hhZG93VG9waWMSNi5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVxdWVzdBo3LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93VG9waWNSZXNwb25zZSLtAZJBnwESEEdldCBTaGFkb3cgVG9waWMaQVJldHJpZXZlIGRldGFpbHMgb2YgYSBzcGVjaWZpYyBzaGFkb3cgdG9waWMgd2l0aGluIGEgc2hhZG93IGxpbmsuSkgKAzIwMBJBCgJPSxI7CjkaNy5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd1RvcGljUmVzcG9uc2WKph0ECAEQA4LT5JMCPBI6L3YxYWxwaGEzL3NoYWRvd2xpbmsve3NoYWRvd19saW5rX25hbWV9L3RvcGljL3t0b3BpY19uYW1lfRKdAwoQR2V0U2hhZG93TWV0cmljcxI4LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TWV0cmljc1JlcXVlc3QaOS5yZWRwYW5kYS5hcGkuZGF0YXBsYW5lLnYxYWxwaGEzLkdldFNoYWRvd01ldHJpY3NSZXNwb25zZSKTApJB3AESF0dldCBTaGFkb3cgTGluayBNZXRyaWNzGnVSZXRyaWV2ZSBtZXRyaWNzIGZvciBhIHNwZWNpZmljIHNoYWRvdyBsaW5rLCBpbmNsdWRpbmcgdG90YWwgdG9waWNzIHJlcGxpY2F0ZWQsIGZhaWxlZCBvdmVyIHRvcGljcywgYW5kIGVycm9yIHRvcGljcy5KSgoDMjAwEkMKAk9LEj0KOxo5LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TWV0cmljc1Jlc3BvbnNliqYdBAgBEAOC0+STAiUSIy92MWFscGhhMy9zaGFkb3dsaW5rL3tuYW1lfS9tZXRyaWNzEr8CCg1HZXRTaGFkb3dMaW5rEjUucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhMy5HZXRTaGFkb3dMaW5rUmVxdWVzdBo2LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TGlua1Jlc3BvbnNlIr4BkkGPARIPR2V0IFNoYWRvdyBMaW5rGjNSZXRyaWV2ZSBkZXRhaWxzIG9mIGEgc3BlY2lmaWMgc2hhZG93IGxpbmsgYnkgbmFtZS5KRwoDMjAwEkAKAk9LEjoKOBo2LnJlZHBhbmRhLmFwaS5kYXRhcGxhbmUudjFhbHBoYTMuR2V0U2hhZG93TGlua1Jlc3BvbnNliqYdBAgBEAOC0+STAh0SGy92MWFscGhhMy9zaGFkb3dsaW5rL3tuYW1lfRpOkkFLChhTaGFkb3cgTGlua3MgKERhdGFwbGFuZSkSL0RhdGFwbGFuZSBvcGVyYXRpb25zIGZvciBSZWRwYW5kYSBzaGFkb3cgbGlua3MuQr4CCiNjb20ucmVkcGFuZGEuYXBpLmRhdGFwbGFuZS52MWFscGhhM0IPU2hhZG93bGlua1Byb3RvUAFaZ2dpdGh1Yi5jb20vcmVkcGFuZGEtZGF0YS9jb25zb2xlL2JhY2tlbmQvcGtnL3Byb3RvZ2VuL3JlZHBhbmRhL2FwaS9kYXRhcGxhbmUvdjFhbHBoYTM7ZGF0YXBsYW5ldjFhbHBoYTOiAgNSQUSqAh9SZWRwYW5kYS5BcGkuRGF0YXBsYW5lLlYxYWxwaGEzygIfUmVkcGFuZGFcQXBpXERhdGFwbGFuZVxWMWFscGhhM+ICK1JlZHBhbmRhXEFwaVxEYXRhcGxhbmVcVjFhbHBoYTNcR1BCTWV0YWRhdGHqAiJSZWRwYW5kYTo6QXBpOjpEYXRhcGxhbmU6OlYxYWxwaGEzYgZwcm90bzM", [file_buf_validate_validate, file_google_api_annotations, file_google_api_field_behavior, file_google_protobuf_timestamp, file_protoc_gen_openapiv2_options_annotations, file_redpanda_api_auth_v1_authorization, file_redpanda_core_admin_v2_shadow_link]);
 
 /**
  * @generated from message redpanda.api.dataplane.v1alpha3.FailOverResponse
@@ -309,6 +309,186 @@ export const GetShadowMetricsResponseSchema: GenMessage<GetShadowMetricsResponse
   messageDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 7);
 
 /**
+ * @generated from message redpanda.api.dataplane.v1alpha3.GetShadowLinkRequest
+ */
+export type GetShadowLinkRequest = Message<"redpanda.api.dataplane.v1alpha3.GetShadowLinkRequest"> & {
+  /**
+   * @generated from field: string name = 1;
+   */
+  name: string;
+};
+
+/**
+ * Describes the message redpanda.api.dataplane.v1alpha3.GetShadowLinkRequest.
+ * Use `create(GetShadowLinkRequestSchema)` to create a new message.
+ */
+export const GetShadowLinkRequestSchema: GenMessage<GetShadowLinkRequest> = /*@__PURE__*/
+  messageDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 8);
+
+/**
+ * @generated from message redpanda.api.dataplane.v1alpha3.GetShadowLinkResponse
+ */
+export type GetShadowLinkResponse = Message<"redpanda.api.dataplane.v1alpha3.GetShadowLinkResponse"> & {
+  /**
+   * @generated from field: redpanda.api.dataplane.v1alpha3.ShadowLink shadow_link = 1;
+   */
+  shadowLink?: ShadowLink;
+};
+
+/**
+ * Describes the message redpanda.api.dataplane.v1alpha3.GetShadowLinkResponse.
+ * Use `create(GetShadowLinkResponseSchema)` to create a new message.
+ */
+export const GetShadowLinkResponseSchema: GenMessage<GetShadowLinkResponse> = /*@__PURE__*/
+  messageDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 9);
+
+/**
+ * Status of a task
+ *
+ * @generated from message redpanda.api.dataplane.v1alpha3.ShadowLinkTaskStatus
+ */
+export type ShadowLinkTaskStatus = Message<"redpanda.api.dataplane.v1alpha3.ShadowLinkTaskStatus"> & {
+  /**
+   * Name of the task
+   *
+   * @generated from field: string name = 1;
+   */
+  name: string;
+
+  /**
+   * State of the task
+   *
+   * @generated from field: redpanda.api.dataplane.v1alpha3.TaskState state = 2;
+   */
+  state: TaskState;
+
+  /**
+   * Reason for task being in state
+   *
+   * @generated from field: string reason = 3;
+   */
+  reason: string;
+
+  /**
+   * The broker the task is running on
+   *
+   * @generated from field: int32 broker_id = 4;
+   */
+  brokerId: number;
+
+  /**
+   * The shard the task is running on
+   *
+   * @generated from field: int32 shard_id = 5;
+   */
+  shardId: number;
+};
+
+/**
+ * Describes the message redpanda.api.dataplane.v1alpha3.ShadowLinkTaskStatus.
+ * Use `create(ShadowLinkTaskStatusSchema)` to create a new message.
+ */
+export const ShadowLinkTaskStatusSchema: GenMessage<ShadowLinkTaskStatus> = /*@__PURE__*/
+  messageDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 10);
+
+/**
+ * @generated from message redpanda.api.dataplane.v1alpha3.ShadowLink
+ */
+export type ShadowLink = Message<"redpanda.api.dataplane.v1alpha3.ShadowLink"> & {
+  /**
+   * Return name, configuration, and overall state
+   *
+   * @generated from field: string name = 1;
+   */
+  name: string;
+
+  /**
+   * @generated from field: string uid = 2;
+   */
+  uid: string;
+
+  /**
+   * @generated from field: redpanda.core.admin.v2.ShadowLinkConfigurations configurations = 3;
+   */
+  configurations?: ShadowLinkConfigurations;
+
+  /**
+   * @generated from field: redpanda.core.admin.v2.ShadowLinkState state = 4;
+   */
+  state: ShadowLinkState;
+
+  /**
+   * @generated from field: repeated redpanda.api.dataplane.v1alpha3.ShadowLinkTaskStatus tasks_status = 5;
+   */
+  tasksStatus: ShadowLinkTaskStatus[];
+
+  /**
+   * @generated from field: repeated string synced_shadow_topic_properties = 6;
+   */
+  syncedShadowTopicProperties: string[];
+};
+
+/**
+ * Describes the message redpanda.api.dataplane.v1alpha3.ShadowLink.
+ * Use `create(ShadowLinkSchema)` to create a new message.
+ */
+export const ShadowLinkSchema: GenMessage<ShadowLink> = /*@__PURE__*/
+  messageDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 11);
+
+/**
+ * Task states
+ *
+ * @generated from enum redpanda.api.dataplane.v1alpha3.TaskState
+ */
+export enum TaskState {
+  /**
+   * @generated from enum value: TASK_STATE_UNSPECIFIED = 0;
+   */
+  UNSPECIFIED = 0,
+
+  /**
+   * Task is active
+   *
+   * @generated from enum value: TASK_STATE_ACTIVE = 1;
+   */
+  ACTIVE = 1,
+
+  /**
+   * Task was paused
+   *
+   * @generated from enum value: TASK_STATE_PAUSED = 2;
+   */
+  PAUSED = 2,
+
+  /**
+   * Task is unable to communicate with source cluster
+   *
+   * @generated from enum value: TASK_STATE_LINK_UNAVAILABLE = 3;
+   */
+  LINK_UNAVAILABLE = 3,
+
+  /**
+   * Task is not running
+   *
+   * @generated from enum value: TASK_STATE_NOT_RUNNING = 4;
+   */
+  NOT_RUNNING = 4,
+
+  /**
+   * Task is faulted
+   *
+   * @generated from enum value: TASK_STATE_FAULTED = 5;
+   */
+  FAULTED = 5,
+}
+
+/**
+ * Describes the enum redpanda.api.dataplane.v1alpha3.TaskState.
+ */
+export const TaskStateSchema: GenEnum<TaskState> = /*@__PURE__*/
+  enumDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 0);
+
+/**
  * @generated from service redpanda.api.dataplane.v1alpha3.ShadowLinkService
  */
 export const ShadowLinkService: GenService<{
@@ -343,6 +523,14 @@ export const ShadowLinkService: GenService<{
     methodKind: "unary";
     input: typeof GetShadowMetricsRequestSchema;
     output: typeof GetShadowMetricsResponseSchema;
+  },
+  /**
+   * @generated from rpc redpanda.api.dataplane.v1alpha3.ShadowLinkService.GetShadowLink
+   */
+  getShadowLink: {
+    methodKind: "unary";
+    input: typeof GetShadowLinkRequestSchema;
+    output: typeof GetShadowLinkResponseSchema;
   },
 }> = /*@__PURE__*/
   serviceDesc(file_redpanda_api_dataplane_v1alpha3_shadowlink, 0);
