+++ proto/redpanda/api/dataplane/v1alpha3/shadowlink.proto
@@ -111,6 +111,56 @@ message GetShadowMetricsResponse {
   uint64 error_topics = 3;
 }
 
+message GetShadowLinkRequest {
+  string name = 1 [
+    (google.api.field_behavior) = REQUIRED,
+    (buf.validate.field).required = true
+  ];
+}
+
+message GetShadowLinkResponse {
+  ShadowLink shadow_link = 1;
+}
+
+// Task states
+enum TaskState {
+  TASK_STATE_UNSPECIFIED = 0;
+  // Task is active
+  TASK_STATE_ACTIVE = 1;
+  // Task was paused
+  TASK_STATE_PAUSED = 2;
+  // Task is unable to communicate with source cluster
+  TASK_STATE_LINK_UNAVAILABLE = 3;
+  // Task is not running
+  TASK_STATE_NOT_RUNNING = 4;
+  // Task is faulted
+  TASK_STATE_FAULTED = 5;
+}
+
+// Status of a task
+message ShadowLinkTaskStatus {
+  // Name of the task
+  string name = 1;
+  // State of the task
+  TaskState state = 2;
+  // Reason for task being in state
+  string reason = 3;
+  // The broker the task is running on
+  int32 broker_id = 4;
+  // The shard the task is running on
+  int32 shard_id = 5;
+}
+
+message ShadowLink {
+  // Return name, configuration, and overall state
+  string name = 1;
+  string uid = 2;
+  core.admin.v2.ShadowLinkConfigurations configurations = 3;
+  core.admin.v2.ShadowLinkState state = 4;
+  repeated ShadowLinkTaskStatus tasks_status = 5;
+  repeated string synced_shadow_topic_properties = 6;
+}
+
 service ShadowLinkService {
   option (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_tag) = {
     name: "Shadow Links (Dataplane)"
@@ -194,6 +244,27 @@ service ShadowLinkService {
           description: "OK"
           schema: {
             json_schema: {ref: ".redpanda.api.dataplane.v1alpha3.GetShadowMetricsResponse"}
+          }
+        }
+      }
+    };
+    option (redpanda.api.auth.v1.authorization) = {
+      required_permission: PERMISSION_VIEW
+      api: API_REDPANDA_ADMIN
+    };
+  }
+
+  rpc GetShadowLink(GetShadowLinkRequest) returns (GetShadowLinkResponse) {
+    option (google.api.http) = {get: "/v1alpha3/shadowlink/{name}"};
+    option (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_operation) = {
+      summary: "Get Shadow Link"
+      description: "Retrieve details of a specific shadow link by name."
+      responses: {
+        key: "200"
+        value: {
+          description: "OK"
+          schema: {
+            json_schema: {ref: ".redpanda.api.dataplane.v1alpha3.GetShadowLinkResponse"}
           }
         }
       }
